#10<Feat>: jerry small refactor
